backend/requirements.txt
Flask
pytesseract
Pillow
spacy
gunicorn
